Update Buffer with validator request done

+++ routes/api.php
@@ -62,16 +62,22 @@ Route::prefix('admin')->group(function () {
     //?ROUTES EN POST
     Route::prefix('POST')->group(function () {
         //VALIDATE BUFFER --> ACTOR
-        Route::post('validate', 'ActorController@store')->name('validate.store');
+        Route::post('validate/{id}', 'ActorController@store')->name('validate.store');
     });
 
-    //?ROUTES EN POST
+    //?ROUTES EN PUT
+    Route::prefix('PUT')->group(function () {
+        //MODIFIER BUFFER
+        Route::put('buffer/{id}', 'BufferController@update')->name('buffer.update');
+    });
+
+    //?ROUTES EN DELETE
     Route::prefix('DELETE')->group(function () {
 
         // SUPPRIMER ACTOR
         Route::delete('actor/{id}', 'ActorController@destroy')->name('actor.destroy');
 
         // SUPPRIMER BUFFER
-        Route::delete('buffer/{id}', 'BufferController@destroy')->name('buffer.destroy');
+        Route::delete('buffer/{id}', 'BufferController@destroy')->name('buffer.destroy')->middleware('auth:admin');
     });
 });
